+++ lib/src/bones_api_entity.dart
@@ -4725,7 +4725,8 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
 
     if (error != null) {
       stackTrace ??= StackTrace.current;
-      _onExecutionError(error, stackTrace, null, null, rethrowError: false);
+      _onExecutionError(error, stackTrace, null, null, null,
+          rethrowError: false);
     }
 
     return asyncTry(commit, onFinally: () => _close(zone: zone));
@@ -4734,21 +4735,23 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
   final List<FutureOr> _executionsFutures = <FutureOr>[];
 
   FutureOr<R> addExecution<R, C>(TransactionExecution<R, C> exec,
-      {Object? Function(Object error, StackTrace stackTrace)? errorResolver,
+      {Object? Function(Object error, StackTrace stackTrace, Object? operation)?
+          errorResolver,
+      Object? operation,
       String? Function()? debugInfo}) {
     if (isFinished) {
       throw StateError("Transaction already finished:\n$this");
     }
 
     if (_executionsFutures.isEmpty) {
-      var ret = _executeSafe(exec, errorResolver, debugInfo);
+      var ret = _executeSafe(exec, operation, errorResolver, debugInfo);
       _executionsFutures.add(ret);
       return ret;
     } else {
       var last = _executionsFutures.last;
 
       var ret = last.resolveWith(() {
-        return _executeSafe(exec, errorResolver, debugInfo);
+        return _executeSafe(exec, operation, errorResolver, debugInfo);
       });
 
       _executionsFutures.add(ret);
@@ -4758,26 +4761,31 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
 
   FutureOr<R> _executeSafe<R, C>(
       TransactionExecution<R, C> exec,
-      Object? Function(Object error, StackTrace stackTrace)? errorResolver,
+      Object? operation,
+      Object? Function(Object error, StackTrace stackTrace, Object? operation)?
+          errorResolver,
       String? Function()? debugInfo) {
     try {
       var ret = exec(context! as C);
       if (ret is Future<R>) {
         var future = ret;
-        return future.catchError(
-            (e, s) => _onExecutionError<R>(e, s, errorResolver, debugInfo));
+        return future.catchError((e, s) =>
+            _onExecutionError<R>(e, s, operation, errorResolver, debugInfo));
       } else {
         return ret;
       }
     } catch (e, s) {
-      return _onExecutionError<R>(e, s, errorResolver, debugInfo);
+      return _onExecutionError<R>(e, s, operation, errorResolver, debugInfo);
     }
   }
 
   FutureOr<R> notifyExecutionError<R>(Object error, StackTrace stackTrace,
-      {Object? Function(Object error, StackTrace stackTrace)? errorResolver,
+      {Object? Function(Object error, StackTrace stackTrace, Object? operation)?
+          errorResolver,
+      Object? operation,
       String? Function()? debugInfo}) {
-    return _onExecutionError<R>(error, stackTrace, errorResolver, debugInfo);
+    return _onExecutionError<R>(
+        error, stackTrace, operation, errorResolver, debugInfo);
   }
 
   Object? _error;
@@ -4787,13 +4795,15 @@ class Transaction extends JsonEntityCacheSimple implements EntityProvider {
   FutureOr<R> _onExecutionError<R>(
       Object error,
       StackTrace stackTrace,
-      Object? Function(Object error, StackTrace stackTrace)? errorResolver,
+      Object? operation,
+      Object? Function(Object error, StackTrace stackTrace, Object? operation)?
+          errorResolver,
       String? Function()? debugInfo,
       {bool rethrowError = true}) {
     var info = debugInfo != null ? debugInfo() : null;
 
     if (errorResolver != null) {
-      error = errorResolver(error, stackTrace) ?? error;
+      error = errorResolver(error, stackTrace, operation) ?? error;
     }
 
     _errorsTransactions[error] = this;