+++ lib/src/bones_api_entity_db_object_directory.dart
@@ -683,9 +683,11 @@ class DBObjectDirectoryAdapter
     return entityJsonNormalized;
   }
 
-  Object resolveError(Object error, StackTrace stackTrace) =>
+  Object resolveError(Object error, StackTrace stackTrace, Object? operation) =>
       DBObjectDirectoryAdapterException('error', '$error',
-          parentError: error, parentStackTrace: stackTrace);
+          parentError: error,
+          parentStackTrace: stackTrace,
+          operation: operation);
 
   FutureOr<R> _finishOperation<T, R>(
       TransactionOperation op, T res, PreFinishDBOperation<T, R>? preFinish) {
@@ -706,6 +708,7 @@ class DBObjectDirectoryAdapter
                 e,
                 s,
                 errorResolver: resolveError,
+                operation: op,
                 debugInfo: () => op.toString(),
               ));
     }
@@ -714,6 +717,7 @@ class DBObjectDirectoryAdapter
       return transaction.addExecution<R, DBObjectDirectoryAdapterContext>(
         (c) => f(c),
         errorResolver: resolveError,
+        operation: op,
         debugInfo: () => op.toString(),
       );
     }
@@ -732,6 +736,7 @@ class DBObjectDirectoryAdapter
       return transaction.addExecution<R, DBObjectDirectoryAdapterContext>(
         (c) => f(c),
         errorResolver: resolveError,
+        operation: op,
         debugInfo: () => op.toString(),
       );
     });
@@ -744,7 +749,6 @@ class DBObjectDirectoryAdapterException extends DBObjectAdapterException {
   String get runtimeTypeNameSafe => 'DBObjectDirectoryAdapterException';
 
   DBObjectDirectoryAdapterException(String type, String message,
-      {Object? parentError, StackTrace? parentStackTrace})
-      : super(type, message,
-            parentError: parentError, parentStackTrace: parentStackTrace);
+      {super.parentError, super.parentStackTrace, super.operation})
+      : super(type, message);
 }