+++ lib/src/bones_api_entity_db_sql.dart
@@ -5,6 +5,7 @@ import 'package:reflection_factory/reflection_factory.dart';
 import 'package:statistics/statistics.dart';
 import 'package:swiss_knife/swiss_knife.dart' show parseBool;
 
+import 'bones_api_base.dart';
 import 'bones_api_condition.dart';
 import 'bones_api_condition_encoder.dart';
 import 'bones_api_condition_sql.dart';
@@ -16,13 +17,14 @@ import 'bones_api_entity_db_relational.dart';
 import 'bones_api_entity_reference.dart';
 import 'bones_api_extension.dart';
 import 'bones_api_initializable.dart';
+import 'bones_api_logging.dart';
 import 'bones_api_platform.dart';
 import 'bones_api_sql_builder.dart';
 import 'bones_api_types.dart';
 import 'bones_api_utils_collections.dart';
 import 'bones_api_utils_json.dart';
 
-final _log = logging.Logger('SQLAdapter');
+final _log = logging.Logger('SQLAdapter')..registerAsSqlLogger();
 
 /// [SQL] wrapper interface.
 abstract class SQLWrapper {
@@ -262,6 +264,16 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
     _boot = true;
 
     DBSQLMemoryAdapter.boot();
+    _setupSQLLogger();
+  }
+
+  static _setupSQLLogger() {
+    var apiRoot = APIRoot.get(singleton: false);
+    var apiConfig = apiRoot?.apiConfig;
+    var logSqlDestiny = apiConfig?.getPath('log', 'sql');
+    if (logSqlDestiny != null) {
+      logSQLTo(logDestiny: logSqlDestiny);
+    }
   }
 
   static final DBAdapterRegister<Object, DBSQLAdapter<Object>> adapterRegister =
@@ -1595,9 +1607,11 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
     });
   }
 
-  Object resolveError(Object error, StackTrace stackTrace) =>
+  Object resolveError(Object error, StackTrace stackTrace, Object? operation) =>
       DBSQLAdapterException('error', '$error',
-          parentError: error, parentStackTrace: stackTrace);
+          parentError: error,
+          parentStackTrace: stackTrace,
+          operation: operation);
 
   @override
   bool isTransactionWithSingleOperation(TransactionOperation op,
@@ -1616,6 +1630,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
                 e,
                 s,
                 errorResolver: resolveError,
+                operation: op,
                 debugInfo: () => sql.mainSQL.toString(),
               ));
     }
@@ -1624,6 +1639,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
       return transaction.addExecution<R, C>(
         f,
         errorResolver: resolveError,
+        operation: op,
         debugInfo: () => sql.mainSQL.toString(),
       );
     }
@@ -1641,6 +1657,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
       return transaction.addExecution<R, C>(
         f,
         errorResolver: resolveError,
+        operation: op,
         debugInfo: () => sql.mainSQL.toString(),
       );
     });
@@ -2005,6 +2022,14 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
     }
   }
 
+  FutureOr<R> _finishSQLOperation<T, R>(SQL sql, TransactionOperation op, T r,
+      PreFinishDBOperation<T, R>? preFinish) {
+    var sqlTime = DateTime.now().difference(op.initTime);
+    var msg = '[time: ${sqlTime.inMilliseconds} ms] $sql';
+    _log.log(logging.Level.INFO, msg);
+    return _finishOperation<T, R>(op, r, preFinish);
+  }
+
   @override
   FutureOr<R> doDelete<R>(TransactionOperation op, String entityName,
       String table, EntityMatcher matcher,
@@ -2030,7 +2055,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
     return generateInsertSQL(op.transaction, entityName, table, fields)
         .resolveMapped((sql) {
       return insertSQL(op, entityName, table, sql, fields)
-          .resolveMapped((r) => _finishOperation(op, r, preFinish));
+          .resolveMapped((r) => _finishSQLOperation(sql, op, r, preFinish));
     });
   }
 
@@ -2059,8 +2084,8 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
       {PreFinishDBOperation<Map<String, dynamic>?, R?>? preFinish}) {
     return generateSelectSQL(op.transaction, entityName, table, ConditionID(id))
         .resolveMapped((sql) {
-      return selectSQL(op, entityName, table, sql)
-          .resolveMapped((r) => _finishOperation(op, r.firstOrNull, preFinish));
+      return selectSQL(op, entityName, table, sql).resolveMapped(
+          (r) => _finishSQLOperation(sql, op, r.firstOrNull, preFinish));
     });
   }
 
@@ -2073,7 +2098,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
             op.transaction, entityName, table, ConditionIdIN(ids))
         .resolveMapped((sql) {
       return selectSQL(op, entityName, table, sql)
-          .resolveMapped((r) => _finishOperation(op, r, preFinish));
+          .resolveMapped((r) => _finishSQLOperation(sql, op, r, preFinish));
     });
   }
 
@@ -2085,7 +2110,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
     return generateSelectSQL(op.transaction, entityName, table, ConditionANY())
         .resolveMapped((sql) {
       return selectSQL(op, entityName, table, sql)
-          .resolveMapped((r) => _finishOperation(op, r, preFinish));
+          .resolveMapped((r) => _finishSQLOperation(sql, op, r, preFinish));
     });
   }
 
@@ -2104,7 +2129,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
             limit: limit)
         .resolveMapped((sql) {
       return selectSQL(op, entityName, table, sql)
-          .resolveMapped((r) => _finishOperation(op, r, preFinish));
+          .resolveMapped((r) => _finishSQLOperation(sql, op, r, preFinish));
     });
   }
 
@@ -2397,7 +2422,6 @@ class DBSQLAdapterException extends DBAdapterException {
   String get runtimeTypeNameSafe => 'DBSQLAdapterException';
 
   DBSQLAdapterException(String type, String message,
-      {Object? parentError, StackTrace? parentStackTrace})
-      : super(type, message,
-            parentError: parentError, parentStackTrace: parentStackTrace);
+      {super.parentError, super.parentStackTrace, super.operation})
+      : super(type, message);
 }